Implement state machine for RSC encoder/decoder
Added rsc.rs with a StateMachine struct to handle both encoder and
decoder functionality for a single input bit. The State newtype is
meant to represent the state of the state machine (for both
encoder and decoder). The BranchMetric struct has fields to
represent different components of the branch metric for a single
state transition (for use in the decoder).

The StateMachine struct has methods to generate output bits and
update state for a given input bit, and to compute the branch
metric for a single state transition.