Implement beta calculator for BCJR decoder
The rsc::BetaCalculator struct has memory for the beta values of
all states after each information bit. Its most significant method
updates the beta value for a state at the previous time instant,
given the state at the current time instant and the branch metric
for the state transition.

The beta values for all states are recentered at each time instant
so that state 0 always has a beta value of 0. This is to prevent
the values from becoming too large (although this might not be a
real concern in a floating-point implementation).

The struct stores the decoding algorithm as a field, but there is
really no need to do so (this is just for convenience).