Implement alpha calculator for BCJR decoder
The rsc::AlphaCalculator struct needs to store the alpha values of
all states only for one time instant (in contrast to the beta
calculator). Its most significant method updates the alpha value
of a state at the next time instant, given the state at the
current time instant and the branch metric for the state
transition.

As with the beta values, the alpha values are recentered so that
state 0 always has an alpha value of 0.

Again, storing the decoding algorithm as a field is probably
unnecessary.