Add function for RSC decoding (BCJR algorithm)
The rsc::decode function takes a mutable reference to a decoder
workspace, and writes the extrinsic information and posterior LLR
values to it.

There are separate helper functions for the BCJR backward and
forward passes, and these in turn rely on helper functions to
compute beta (resp. alpha) values for all states at the previous
(resp. next) time instant.