Implement interleaver and deinterleaver
Both are implemented within a single Interleaver struct, which is
generic over the type of the elements of the sequence. They are
specified by a permutation of the integers 0, 1, ..., L-1, where
L is the length of the sequence.

The interleave and deinterleave methods both take mutable
references to a buffer to write their respective outputs (so that
no memory allocation is needed within them).