Strengthen test of BCJR forward pass
Added helper functions to return the correct extrinsic information
and posterior LLRs for each of 4 information bits (to allow easy
testing with different values of code bit LLRs and prior LLRs for
the information bits).

Also removed the call to rsc::run_bcjr_backward_pass in this test,
and instead directly set all required beta values in the workspace
(so as to isolate the function of interest in the test).