Fix bug in constructing top/bottom BCJR inputs
The error was to interleave the parity bit LLR values intended for
the bottom BCJR decoder. Only the systematic bit LLR values should
have been interleaved (needed to replace i_bottom with i_top in the
bcjr_inputs function to fix this error). Unfortunately, the same
error was also present in the test for the bcjr_inputs function, so
it was not caught earlier (the test is also now fixed).

This error was introduced in the initial implementation of the PCCC
decoder, in commit 075cafe. With this fix, everything seems to be
working fine!