Improve crate documentation
Added a heading and some italics in the top-level documentation.
Clarified what "second" encoder means. Replaced BCJR by APP in the
description of the decoder (should make this change throughout).
Changed the parameters of the example (different choice of 4 info
bits and interleaver), just so that more states are visited. Also
made all the code bit LLR values of magnitude 1 (instead of 10),
and changed the decoding algorithm to Log-MAP (which is probably
better known than Linear-Log-MAP). Added some clarifying comments
in the example.

In the documentation for the encoder and decoder functions, added
some text to clarify what it means for the code polynomials to be
invalid, and changed the examples to match the top-level crate
documentation.