Make JSON read/write functions private
These are in the lte module, and there's no need to have them in
the public interface. The read function is currently unused, which
raises a clippy lint, but the goal is to use it in a function for
plotting simulation results (to be implemented).

Also slipped in a very minor docstring edit.