Add SimCase struct in lte module
This struct is meant to capture all the parameters that make up a
distinct simulation case, in plots of BLER or BER versus SNR. For
now, a simulation case is described by the block size and the
decoding algorithm.

Derived the Hash trait for SimCase (which also required deriving
Hash for DecodingAlgo). This will be needed in order to find all
the unique simulation cases represented in a JSON file.