Implement a couple of traits for lte::SimCase
Implemented Display, making use of the Display implementation for
DecodingAlgo. Also implemented From<&SimResults>, meant to extract
the simulation case corresponding to given simulation results. Did
this for a reference type to avoid consuming the SimResults object
(even though SimResults is Copy).

Also moved the declaration of SimCase, which was originally placed
by mistake between the encoder and decoder functions.