Make breaking change in command line parser
Summary of changes:
1. Now only one block size is accepted. There was no good reason to
allow multiple block sizes, and it is cleaner to have just the SNR
take on multiple values.
2. Renamed the two parameters for the decoding algorithm (name and
number of iterations), to make it clearer that they together
constitute the decoding algorithm.
3. Incorporated the helper function for getting the number of
decoding iterations from the command-line matches directly into
the function to return the decoding algorithm.
4. Changed various default values and edited various comments for
clarity and consistency.