Make print statements in lte module more concise
The simulation parameters are being printed out anyway, so there is
no need to repeat the block size and SNR on the progress line. Also,
the BER and BLER are now shown just as ratios of integers. The final
BER and BLER are printed at the end in the summary.