Rename lte::summarize_all_sim_results function
Added _in_file at the end to indicate that it reads simulation
results from a file. Also added a helper function to do the actual
parsing and printing. Removed the call to the summarize function
from main.rs and instead added a call to the above helper function
in lte::run_bpsk_awgn_sims.

Removed all the tests of private functions that create or read from
a JSON file. Retained just the test of lte::run_bpsk_awgn_sims, but
marked it "ignore", so that it can be run separately when needed.