Save results to file after each simulation case
In lte::run_bpsk_awgn_sims, moved the call to the function saving
all simulation results to file inside the loop over all cases.
This means the whole set of results up to that point is saved
again after each case is completed. This is a bit redundant, but
it is probably better than saving everything only at the end.

What would be best is saving incrementally to the file after each
case is completed.