Make yet more README edits
Now there's only one top-level heading, and it's now called pccc
(the markdown linter was complaining about multiple top-level
headings).

Used "cargo run" in the hint about how to run the command-line
executable (earlier, the path to the executable was directly
specified).

Removed the leading dollar symbol in the console output shown in
the example section (otherwise, the linter was complaining that
no output was shown).

More obsessive and pointless rewording, etc.