Speed up LTE BPSK-AWGN simulations with rayon
Refactored lte::run_bpsk_awgn_sims to use par_iter from rayon.
This results in 6-7x speedup when simulating 1000 blocks at 11
different SNR values, with a block size of 1536 bits.

However, to use rayon, the thread-local random number generator
is directly accessed in lte::bpsk_awgn_sim (which is accordingly
also refactored). This means there is no longer a `rng` argument
in these two methods. Unfortunately, this means reproducibility
is no longer assured.

Further, the simulation parameters and the progress message during
the simulation are no longer printed. The reason is that these
print statements get all mixed up due to the parallel iteration,
so the output is not meaningful.

In the doctest for lte::run_bpsk_awgn_sims, uncommented the call
to the actual function (which means a results.json file will be
created after the doctest is run). Changed the parameters so that
the doctest runs fast.