Speed up lte::bpsk_awgn_sim with rayon
Refactored this function so that the check for the simulation
being completed is done only at the end of a run of blocks. Used
into_par_iter from rayon to parallelize the simulation of each
such run of blocks. This leads to a very significant speed-up.

The downside (for now, at least) is that the random_bits and
bpsk_awgn_channel functions in the utils module no longer take a
random number generator as argument (instead, the thread-local
rng is called in these functions). But the speed-up is worth the
loss in exact reproducibility.

For consistency, removed the rng argument from Interleaver::random
also. Now there are no explicit rng arguments anywhere.