Remove lte::summarize_all_sim_results_in_file
This was in the public interface, so this is a breaking change.
The reason for removing the function is that it was not sound in
the first place. Any JSON file could be passed to it, and there
was no check that the file came from a simulation run.

The all_sim_results_from_file function is now unused, but it is
retained anyway with a directive to allow dead code.