Add SimParams::print_concise method
This is for use in lte::bpsk_awgn_sim. To identify each simulation
case, it is necessary to print only the block size, the decoding
algorithm and number of turbo iterations, and the Es/N0. Other
simulation parameters are suppressed.

Also added a formatting specification in print_es_over_n0_db, to
have just two decimal places printed.