Remove par_iter from run_bpsk_awgn_sims
Since there is already a par_iter in bpsk_awgn_sim, it is not
clear if any further speed-up is obtained by having a par_iter
in run_bpsk_awgn_sims as well. Some testing suggested there was
no meaningful difference in run time either way. The benefit of
having a regular iter in run_bpsk_awgn_sims is that the
simulations for different scenarios are executed in order.